Update to OpenBSD 7.3
This change breaks the script because OpenBSD 7.0 and later ignore the
TFTP server name. Instead, OpenBSD tries to fetch the installation
configuration from the next-server of the DHCP lease, in this case
10.0.2.2.

I failed to adapt the qemu guest forwarding rule accordingly. Thus, the
script is broken for now, unless you are willing to serve the mirror on
the privileged port 80.